deployment process works

+++ app/Git.php
@@ -8,6 +8,7 @@ class Git
 {
     protected $connection;
     protected $responses;
+    protected $serverDate;
 
     /**
      * @var array
@@ -56,23 +57,19 @@ EOF;
         return $this->responses;
     }
 
-    public function deploy($deployment){
-        $this->server = $deployment->server;
-//        $project = $this->server->project()->first();
-        $commitRef = "7fd1a60b01f91b314f59955a4e4d4e80d8edf11d"; //deleteme
+    protected function getServerDate(){
+        $location = $this->server->deploy_location;
+        $this->responses[] = $res = $this->connection->execute("date +%F_%H-%M-%S");
+        $this->serverDate = str_replace_last("\n",'',$res['message']);
+        $cmd = "cd {$location} && mkdir -p releases/{$this->serverDate} && mkdir shared && echo folders created";
+        $this->responses[] = $this->connection->execute($cmd);
+    }
+
+    protected function cloneAndCheckout(){
         $location = $this->server->deploy_location;
         $repository = $this->server->project()->first()->repository;
+//        $commitRef = "7fd1a60b01f91b314f59955a4e4d4e80d8edf11d"; //deleteme
 
-        if ($this->testSshConnection() === false) {
-            return ['output' => $this->responses, 'success' => false];
-        }
-
-        $this->updateCache();
-
-        $this->responses[] = $res = $this->connection->execute("date +%F_%H-%M-%S");
-        $date = str_replace_last("\n",'',$res['message']);
-        $cmd = "cd {$location} && mkdir -p releases/{$date} && mkdir shared && echo folders created";
-        $this->responses[] = $this->connection->execute($cmd);
         $cmd = <<<'EOF'
         function deployGit() {
             local repositoryUrl="${1}";
@@ -85,19 +82,32 @@ EOF;
             echo git clone created;
         }
 EOF;
-        $releaseLocation = "{$location}/releases/{$date}";
+        $releaseLocation = "{$location}/releases/{$this->serverDate}";
         $refFolder = "${location}/.gitcache/".preg_replace("/[^a-zA-Z0-9]/", "-", $repository);
         $cmd .= "\n deployGit $repository $refFolder $commitRef $releaseLocation";
         $this->responses[] = $this->connection->execute($cmd);
+    }
+
+    public function deploy($deployment){
+        $this->server = $deployment->server;
+        $location = $this->server->deploy_location;
+
+        if ($this->testSshConnection() === false) {
+            return ['output' => $this->responses, 'success' => false];
+        }
+
+        $this->updateCache();
+        $this->getServerDate();
+        $this->cloneAndCheckout();
 
         $this->responses[] = $this->connection->execute("###### links files from shared ######");
         $this->responses[] = $this->connection->execute("###### do their cmds ######");
         $this->responses[] = $this->connection->execute("cd $location && rm previous ");
         $this->responses[] = $this->connection->execute("cd $location && mv current previous");
-        $this->responses[] = $this->connection->execute("cd $location && ln -s $location/releases/$date current");
+        $this->responses[] = $this->connection->execute("cd $location && ln -s $location/releases/$this->serverDate current");
 
         $this->connection->disconnect();
 
-        return $this->responses;
+        return ['output' => $this->responses, 'success' => true];
     }
 }
